@@ -1,13 +1,16 @@
-package com.cybersoft.cozastore;
+package com.cybersoft.cozaStore;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
-public class CozastoreApplication {
+
+public class CozaStoreApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(CozastoreApplication.class, args);
+
+		SpringApplication.run(CozaStoreApplication.class, args);
 	}
 
 }
