@@ -1,12 +1,16 @@
-package com.cybersoft.cozastore.repository;
+package com.cybersoft.cozaStore.repository;
 
-import com.cybersoft.cozastore.entity.CartEntity;
+import com.cybersoft.cozaStore.entity.CartEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface CartRepository extends JpaRepository<CartEntity, Integer> {
+public interface CartRepository extends JpaRepository<CartEntity,Integer> {
 
-
+    Optional<CartEntity> findByUserIdAndProductId(@Param("id_user") int userId, @Param("id_product") int productId);
 
 }
