@@ -1,18 +1,19 @@
-package com.cybersoft.cozastore.service;
+package com.cybersoft.cozaStore.service;
 
-import com.cybersoft.cozastore.entity.CartEntity;
-import com.cybersoft.cozastore.entity.ProductEntity;
-import com.cybersoft.cozastore.entity.UserEntity;
-import com.cybersoft.cozastore.payload.request.CartRequest;
-import com.cybersoft.cozastore.payload.response.CartResponse;
-import com.cybersoft.cozastore.repository.CartRepository;
-import com.cybersoft.cozastore.repository.ProductRepository;
-import com.cybersoft.cozastore.repository.UserRepository;
-import com.cybersoft.cozastore.service.imp.CartServiceImp;
+import com.cybersoft.cozaStore.entity.CartEntity;
+import com.cybersoft.cozaStore.entity.ProductEntity;
+import com.cybersoft.cozaStore.entity.UserEntity;
+import com.cybersoft.cozaStore.payload.request.CartRequest;
+import com.cybersoft.cozaStore.payload.response.CartResponse;
+import com.cybersoft.cozaStore.repository.CartRepository;
+import com.cybersoft.cozaStore.repository.ProductRepository;
+import com.cybersoft.cozaStore.repository.UserRepository;
+import com.cybersoft.cozaStore.service.imp.CartServiceImp;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
@@ -31,16 +32,28 @@ public class CartService implements CartServiceImp {
     @Override
     public boolean insertProductIntoCart(CartRequest cartRequest) {
 
-        CartEntity cart = new CartEntity();
-
         Optional<ProductEntity> product = productRepository.findById(cartRequest.getIdProduct());
         Optional<UserEntity> user = userRepository.findById(cartRequest.getIdUser());
 
         if(product.isPresent() && user.isPresent()){
 
-            cart.setProduct(product.get());
-            cart.setUser(user.get());
-            cart.setQuanity(cartRequest.getQuanity());
+            CartEntity cart;
+            // Kiểm tra xem sản phẩm đã tồn tại trong giỏ hàng của người dùng hay chưa
+            Optional<CartEntity> existingCartItem = cartRepository.findByUserIdAndProductId(user.get().getId(), product.get().getId());
+
+            if (existingCartItem.isPresent()) {
+                // Nếu sản phẩm đã tồn tại, tăng số lượng
+                cart = existingCartItem.get();
+                cart.setQuanity(cart.getQuanity() + cartRequest.getQuanity());
+
+            }else {
+                // Nếu sản phẩm chưa tồn tại, tạo một mục mới
+                cart = new CartEntity();
+                cart.setProduct(product.get());
+                cart.setUser(user.get());
+                cart.setQuanity(cartRequest.getQuanity());
+                cart.setCreateDate(new Date());
+            }
 
             try {
                 cartRepository.save(cart);
@@ -56,21 +69,86 @@ public class CartService implements CartServiceImp {
     }
 
     @Override
-    public List<CartResponse> getCart(int idUser) {
+    public List<CartResponse> getCartByIdUser(int idUser) {
         List<CartEntity> cartEntities = cartRepository.findAll();
         List<CartResponse> cartResponses = new ArrayList<>();
-        for (CartEntity c: cartEntities){
-            if (c.getUser().getId() == idUser){
+
+        for (CartEntity c : cartEntities) {
+            if (c.getUser().getId() == idUser) {
                 CartResponse cartTemp = new CartResponse();
-                cartTemp.setCart(c.getId());
+                cartTemp.setIdCart(c.getId());
                 cartTemp.setQuanity(c.getQuanity());
                 cartTemp.setNameProduct(c.getProduct().getName());
+                cartTemp.setNameUser(c.getUser().getUsername());
+
+                // Calculate the total price
+//                double totalPrice = c.getProduct().getPrice() * c.getQuanity();
+                cartTemp.setTotalCost(c.getProduct().getPrice() * c.getQuanity());
 
                 cartResponses.add(cartTemp);
-
             }
         }
 
         return cartResponses;
+    }
+
+
+
+    @Override
+    public boolean deleteProductById(int idProduct) {
+        if (productRepository.existsById(idProduct)) {
+            productRepository.deleteById(idProduct); // Xóa sản phẩm
+            return true; // Trả về true nếu xóa thành công
+        } else {
+            return false; // Trả về false nếu không tìm thấy sản phẩm để xóa
+        }}
+
+    @Override
+    public boolean deleteCartByIdUser(int idUser) {
+        if(cartRepository.existsById(idUser)){
+            cartRepository.deleteById(idUser);
+            return true;
+        }else{
+            return false;
+        }
+    }
+
+    @Override
+    public boolean deleteCartById(int id) {
+        if(cartRepository.existsById(id)){
+            cartRepository.deleteById(id);
+            return true;
+        }else{
+            return false;
+        }
+    }
+
+    @Override
+    public boolean updateCart(int id, int id_product, int quanity, int id_user) {
+    Optional<CartEntity> optionalCart = cartRepository.findById(id);
+    List<CartResponse> cartResponses = new ArrayList<>();
+
+    if(optionalCart.isPresent()){
+        CartEntity cartEntity = optionalCart.get();
+        cartEntity.setId(id);
+        cartEntity.setQuanity(quanity);
+        cartEntity.setCreateDate(new Date());
+
+        UserEntity userEntity = new UserEntity();
+        userEntity.setId(id_user);
+        cartEntity.setUser(userEntity);
+
+        ProductEntity productEntity = new ProductEntity();
+        productEntity.setId(id_product);
+        cartEntity.setProduct(productEntity);
+
+        cartRepository.save(cartEntity);
+
+        return true;
+    }else{
+        return false;
+    }
+
+
     }
 }
