@@ -1,6 +1,5 @@
-package com.cybersoft.cozastore.service.imp;
+package com.cybersoft.cozaStore.service.imp;
 
 public interface BlogServiceImp {
-
     boolean insertBlog();
 }
